Added Dash-cryptocurrency, moved Cryptocurrency enum, added explanatory comment
