+++ anydex/wallet/cryptocurrency.py
@@ -1,0 +1,17 @@
+from enum import Enum
+
+
+class Cryptocurrency(Enum):
+    """
+    Enum representing curerntly implemented cryptocurrencies.
+    """
+
+    BITCOIN = 'bitcoin'
+    DASH = 'dash'
+    BANDWIDTH_TOKEN = 'bandwidth_token'
+    ETHEREUM = 'ethereum'
+    RIPPLE = 'ripple'
+    LITECOIN = 'litecoin'
+    IOTA = 'iota'
+    MONERO = 'monero'
+    ZCASH = 'zcash'
